Create a dedicated module for package installers
Create a dedicated module for package installers with the intention
of decoupling installer implementations from the rest of the mod
manager codebase as best as possible.

If decoupling can be achieved to a significant enough degree, it will
help with the transition to CLI-backed mod management and opens the
possibility to easily write tests for installers.
