+++ src/r2mm/installing/profile_installers/GenericProfileInstaller.ts
@@ -18,84 +18,9 @@ import FileWriteError from '../../../model/errors/FileWriteError';
 import FileUtils from '../../../utils/FileUtils';
 import { PackageLoader } from '../../../model/installing/PackageLoader';
 import ZipProvider from "../../../providers/generic/zip/ZipProvider";
-
-const basePackageFiles = ["manifest.json", "readme.md", "icon.png"];
-
-
-
-abstract class PackageInstaller {
-    abstract install(mlLocation: string, modLoaderMapping: ModLoaderPackageMapping, profile: Profile): Promise<void>;
-    // abstract uninstall(): Promise<void>;  // TODO: Implement
-}
-
-
-// TODO: Implement
-// class InstallRuleInstaller extends PackageInstaller {
-//     /**
-//      * The default installer that handles the game-specific install rules
-//      */
-// }
-
-
-class BepInExInstaller extends PackageInstaller {
-    /**
-     * Handles installation of BepInEx
-     */
-    async install(bieLocation: string, modLoaderMapping: ModLoaderPackageMapping, profile: Profile) {
-        let bepInExRoot: string;
-        if (modLoaderMapping.rootFolder.trim().length > 0) {
-            bepInExRoot = path.join(bieLocation, modLoaderMapping.rootFolder);
-        } else {
-            bepInExRoot = path.join(bieLocation);
-        }
-        for (const item of (await FsProvider.instance.readdir(bepInExRoot))) {
-            if (!basePackageFiles.includes(item.toLowerCase())) {
-                if ((await FsProvider.instance.stat(path.join(bepInExRoot, item))).isFile()) {
-                    await FsProvider.instance.copyFile(path.join(bepInExRoot, item), path.join(profile.getPathOfProfile(), item));
-                } else {
-                    await FsProvider.instance.copyFolder(path.join(bepInExRoot, item), path.join(profile.getPathOfProfile(), item));
-                }
-            }
-        }
-    }
-}
-
-
-class MelonLoaderInstaller extends PackageInstaller {
-    /**
-     * Handles installation of MelonLoader
-     */
-    async install(mlLocation: string, modLoaderMapping: ModLoaderPackageMapping, profile: Profile) {
-        for (const item of (await FsProvider.instance.readdir(mlLocation))) {
-            if (!basePackageFiles.includes(item.toLowerCase())) {
-                if ((await FsProvider.instance.stat(path.join(mlLocation, item))).isFile()) {
-                    await FsProvider.instance.copyFile(path.join(mlLocation, item), path.join(profile.getPathOfProfile(), item));
-                } else {
-                    await FsProvider.instance.copyFolder(path.join(mlLocation, item), path.join(profile.getPathOfProfile(), item));
-                }
-            }
-        }
-    }
-}
-
-
-class GodotMLInstaller extends PackageInstaller {
-    /**
-     * Handles installation of GodotML
-     */
-    async install(mlLocation: string, modLoaderMapping: ModLoaderPackageMapping, profile: Profile) {
-        const copyFrom = path.join(mlLocation, "addons", "mod_loader");
-        const copyTo = path.join(profile.getPathOfProfile(), "addons", "mod_loader");
-        const fs = FsProvider.instance;
-
-        if (await fs.exists(copyFrom)) {
-            if (!await fs.exists(copyTo)) {
-                await fs.mkdirs(copyTo);
-            }
-            await fs.copyFolder(copyFrom, copyTo);
-        }
-    }
-}
+import { BepInExInstaller} from "../../../installers/BepInExInstaller";
+import { MelonLoaderInstaller } from "../../../installers/MelonLoaderInstaller";
+import { GodotMLInstaller } from "../../../installers/GodotMLInstaller";
